refactor: use separate folders for lambda functions.
