+++ backend/product_service/lambda/product_by_id/logger.ts
