+++ backend/product_service/lambda/product_by_id/product_by_id.ts
