+++ backend/product_service/lambda/product_create/product_create.ts
