+++ backend/product_service/lambda/product_list/product_list.ts
